% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentencepiece.R
\name{sentencepiece}
\alias{sentencepiece}
\title{Construct a Sentencepiece model}
\usage{
sentencepiece(x, type = c("bpe", "char", "unigram", "word"),
  vocab_size = 8000, coverage = 0.9999,
  model_prefix = "sentencepiece", threads = 1L, args)
}
\arguments{
\item{x}{a character vector of path(s) to the text files containing training data}

\item{type}{either one of 'bpe', 'char', 'unigram' or 'word' for Byte Pair Encoding, Character level encoding,
Unigram encoding or pretokenised word encoding. Defaults to 'bpe' (Byte Pair Encoding).}

\item{vocab_size}{integer indicating the number of tokens in the final vocabulary. Defaults to 8000.}

\item{coverage}{fraction of characters covered by the model. Must be in the range [0, 1]. A good value to use is about 0.9999.}

\item{model_prefix}{character string with the name of the model. Defaults to 'sentencepiece'.
When executing the function 2 files will be created in the current working directory, namely
sentencepiece.model with the model and sentencepiece.vocab containing the vocabulary of the model. 
You can change the name of the model by providing the \code{model_prefix} argument.}

\item{threads}{integer indicating number of threads to use when building the model}

\item{args}{character string with arguments passed on to sentencepiece::SentencePieceTrainer::Train (for expert use only)}
}
\value{
an object of class \code{sentencepiece} which is defined at \code{\link{sentencepiece_load_model}}
}
\description{
Construct a Sentencepiece model on text.
}
\examples{
library(tokenizers.bpe)
data(belgium_parliament, package = "tokenizers.bpe")
writeLines(belgium_parliament$text, con = "traindata.txt")
model <- sentencepiece("traindata.txt", type = "unigram", vocab_size = 5000)
model <- sentencepiece("traindata.txt", type = "char", vocab_size = 5000)
model <- sentencepiece("traindata.txt", type = "bpe", vocab_size = 5000)

txt <- c("De eigendomsoverdracht aan de deelstaten is ingewikkeld.",
         "On est d'accord sur le prix de la biere?")
sentencepiece_encode(model, x = txt, type = "subwords")
sentencepiece_encode(model, x = txt, type = "ids")
model <- sentencepiece_load_model("sentencepiece.model")
sentencepiece_encode(model, x = txt, type = "subwords")
sentencepiece_encode(model, x = txt, type = "ids")


# clean up for CRAN
file.remove("sentencepiece.model")
file.remove("sentencepiece.vocab")
file.remove("traindata.txt")
}
\seealso{
\code{\link{sentencepiece_load_model}}
}
