% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt_remove_}
\alias{txt_remove_}
\title{Remove prefixed underscore}
\usage{
txt_remove_(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
\code{x} where the prefixed underscore is removed
}
\description{
Remove prefixed underscore unicode character \u2581
}
\examples{
x <- c("\u2581word", "hello", "_regularunderscore")
x
txt_remove_(x)
}
