% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpemb.R
\name{read_word2vec}
\alias{read_word2vec}
\title{Read a word2vec embedding file}
\usage{
read_word2vec(x)
}
\arguments{
\item{x}{path to the file}
}
\value{
a matrix with one row per token containing the embedding of the token
}
\description{
Read a word2vec embedding file

Read a word2vec embedding file
}
\examples{
bpe <- sentencepiece_download_model("nl", vocab_size = 1000, dim = 50)
embedding <- read_word2vec(bpe$glove$file_model)
bpe <- sentencepiece_download_model("Dutch", vocab_size = 1000, dim = 50)
embedding <- read_word2vec(bpe$glove$file_model)
bpe <- sentencepiece_download_model("WestFlemish", vocab_size = 50000, dim = 25)
embedding <- read_word2vec(bpe$glove$file_model)
model     <- sentencepiece_load_model(bpe$file_model)
sentencepiece_encode(model, "Mieljaar wuk ne buchtspellement is me daddier")
}
